finish add schedule and list schedule

+++ app/src/main/java/com/fishfeeder/ui/screens/schedule/ScheduleEvent.kt
@@ -2,5 +2,8 @@ package com.fishfeeder.ui.screens.schedule
 
 
 sealed class ScheduleEvent {
-    data class OnChangeSchedule(val id : Int, val isActive : Boolean) : ScheduleEvent()
+
+
+
+    data class OnChangeSchedule(val id : Long, val isActive : Boolean) : ScheduleEvent()
 }
